Moves analysis subpackage into comptools package
