--- comptools/analysis/__init__.py
@@ -1,16 +1,0 @@
-
-__all__ = []
-
-def export(defn):
-    globals()[defn.__name__] = defn
-    __all__.append(defn.__name__)
-    return defn
-
-from .pipelines import *
-from .features import *
-from .data_functions import *
-from .base import get_energybins, get_color_dict
-from .subsample import get_random_subsample
-from .modelevaluation import get_CV_frac_correct
-from .spectrumfunctions import get_flux, get_model_flux
-from .LDFfunctions import fit_DLP_params, DLP
