--- comptools/analysis/base.py
@@ -1,122 +1,0 @@
-
-from collections import namedtuple
-import numpy as np
-
-class DataSet(object):
-
-    def __init__(self, name=None, X=None, y=None, le=None):
-        self.name = name
-        self.X = X  # features array
-        self.y = y  # targets array
-        self.le = le # target LabelEncoder, if exists
-        # If both y and a le are provided, store the
-        # LabelEncoder inverse_transform-ed array as 'labels'
-        if all(item is not None for item in [y, le]):
-            self.labels = le.inverse_transform(y)
-
-    def __repr__(self):
-        output = 'DataSet(name={},\nX={},\ny={})'.format(self.name, self.X, self.y)
-        return output
-
-    def __len__(self):
-        if self.X is not None:
-            return self.X.shape[0]
-        else:
-            return 0
-
-    def __getitem__(self, sliced):
-        sliced_dataset = DataSet(name=self.name, le=self.le)
-        for attr, value in self.__dict__.iteritems():
-            if attr in ['name', 'le']: continue
-            try:
-                sliced_value = value[sliced]
-                setattr(sliced_dataset, attr, sliced_value)
-            except:
-                setattr(sliced_dataset, attr, value)
-
-        return sliced_dataset
-
-    def __add__(self, other):
-        concat_dataset = DataSet(name=self.name, le=self.le)
-        for attr, value in self.__dict__.iteritems():
-            if attr in ['name', 'le']: continue
-            other_value = getattr(other, attr)
-            concat_value = np.concatenate((value, other_value))
-            setattr(concat_dataset, attr, concat_value)
-
-        return concat_dataset
-
-    def isnull(self):
-        attributes = []
-        for attr, value in self.__dict__.iteritems():
-            attributes.append(value)
-        any_attr_true = any(attributes)
-        is_null = not any_attr_true
-
-        return is_null
-
-
-def get_energybins(config='IC86.2012'):
-
-    # Create EnergyBins object to store all realted information
-    EnergyBins = namedtuple('EnergyBins', ['energy_min', 'energy_max',
-        'energy_bins', 'energy_midpoints', 'energy_bin_widths',
-        'log_energy_min', 'log_energy_max', 'log_energy_bin_width',
-        'log_energy_bins', 'log_energy_midpoints'])
-
-    # Define energy range for this analysis
-    if 'IC79' in config:
-        log_energy_min = 6.1
-        log_energy_break = 8.0
-        log_energy_max = 9.0
-        log_energy_bins = np.concatenate(
-                        (np.arange(log_energy_min, log_energy_break-0.1, 0.1),
-                         np.arange(log_energy_break, log_energy_max+0.2, 0.2)))
-    elif 'IC86' in config:
-        log_energy_min = 6.1
-        log_energy_max = 8.0
-        log_energy_bins = np.arange(log_energy_min, log_energy_max+0.1, 0.1)
-    else:
-        raise ValueError(
-            'Invalid detector configuration entered: {}'.format(config))
-
-    log_energy_bin_width = log_energy_bins[1:] - log_energy_bins[:-1]
-    log_energy_midpoints = (log_energy_bins[1:] + log_energy_bins[:-1]) / 2
-
-    energy_min = 10**log_energy_min
-    energy_max = 10**log_energy_max
-    energy_bins = 10**log_energy_bins
-    energy_midpoints = 10**log_energy_midpoints
-    energy_bin_widths = energy_bins[1:] - energy_bins[:-1]
-
-    # Create instance of EnergyBins with appropriate binning
-    energybins = EnergyBins(energy_min=energy_min,
-                            energy_max=energy_max,
-                            energy_bins=energy_bins,
-                            energy_midpoints=energy_midpoints,
-                            energy_bin_widths=energy_bin_widths,
-                            log_energy_min=log_energy_min,
-                            log_energy_max=log_energy_max,
-                            log_energy_bin_width=log_energy_bin_width,
-                            log_energy_bins=log_energy_bins,
-                            log_energy_midpoints=log_energy_midpoints)
-
-    return energybins
-
-
-def get_color_dict():
-    color_dict = {'light': 'C0', 'heavy': 'C1', 'intermediate': 'C3', 'total': 'C2',
-                 'P': 'C0', 'He': 'C1', 'O': 'C3', 'Fe':'C4',
-                 'PPlus': 'C0', 'He4Nucleus': 'C4', 'O16Nucleus': 'C3', 'Fe56Nucleus':'C1',
-                 'data': 'k'}
-
-    return color_dict
-
-def cast_to_ndarray(input):
-    if isinstance(input, np.ndarray):
-        return input
-    else:
-        try:
-            return np.array(input)
-        except:
-            raise TypeError('Input wasn\'t able to be cast to a numpy.ndarray. Got type {}.'.format(type(input)))
