+++ comptools/data_functions.py
@@ -1,27 +1,25 @@
-#!/usr/bin/env python
 
 from __future__ import division
 import numpy as np
 from scipy import stats
-from . import export
 
-@export
+
 def get_summation_error(errors):
     sum_error = np.sqrt(np.sum([err**2 for err in errors]))
     return sum_error
 
-@export
+
 def get_difference_error(errors):
     diff_error = np.sqrt(np.sum([err**2 for err in errors]))
     return diff_error
 
-@export
+
 def get_ratio_error(num, num_err, den, den_err):
     ratio = num/den
     ratio_err = np.absolute(ratio) * np.sqrt((num_err / num)**2 + (den_err / den)**2)
     return ratio_err
 
-@export
+
 def ratio_error(num, num_err, den, den_err, nan_to_num=False):
     ratio = num/den
     ratio_err = np.abs(ratio) * np.sqrt((num_err / num)**2 + (den_err / den)**2)
@@ -29,13 +27,13 @@ def ratio_error(num, num_err, den, den_err, nan_to_num=False):
         ratio_err = np.nan_to_num(ratio_err)
     return ratio, ratio_err
 
-@export
+
 def product_error(term1, term1_err, term2, term2_err):
     product = term1 * term2
     product_err = product * np.sqrt((term1_err / term1)**2 + (term2_err / term2)**2)
     return product, product_err
 
-@export
+
 def averaging_error(values, errors):
     if not isinstance(values, np.ndarray):
         values = np.array(values)
@@ -45,6 +43,7 @@ def averaging_error(values, errors):
     sum_error = np.sum([np.sqrt(error**2) for error in errors])
 
     return sum_value/len(values), sum_error/len(values)
+
 
 def get_bin_mids(bins, infvalue=None):
     abins = np.asarray(bins)
@@ -57,6 +56,7 @@ def get_bin_mids(bins, infvalue=None):
     if abs(steps[-1]) == np.inf:
         mids[-1] = abins[-2] + steps[-2] / 2.
     return mids
+
 
 def get_medians(x, y, bins):
     lower_error = lambda x: np.percentile(x, 16)
@@ -105,6 +105,7 @@ def get_cumprob_sigma(values):
     sigma_containment = bin_midpoints[sigma_index]/1.51
 
     return sigma_containment
+
 
 def get_resolution(x, y, bins):
     binned_statistic, bin_edges, binnumber = stats.binned_statistic(x, y,
