+++ comptools/io.py
